+++ packages/article-topics/__tests__/android/article-topics.test.js
@@ -1,0 +1,5 @@
+import shared from "../shared";
+
+describe("Article Topics test on android: ", () => {
+  shared();
+});