+++ packages/author-profile/src/author-profile-list-item-separator.js
@@ -1,0 +1,9 @@
+import React from "react";
+import { View } from "react-native";
+import styles from "./styles";
+
+const AuthorProfileListItemSeparator = () => (
+  <View style={styles.listItemSeparator} />
+);
+
+export default AuthorProfileListItemSeparator;