+++ packages/topics/topics.stories.js
@@ -1,4 +1,0 @@
-import { showcaseConverter } from "@times-components/storybook";
-import showcase from "./topics.showcase";
-
-showcaseConverter(module, showcase);