+++ packages/author-profile/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.58.10"></a>
+## [0.58.10](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.58.9...@times-components/author-profile@0.58.10) (2018-05-03)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-profile
+
+<a name="0.58.9"></a>
+## [0.58.9](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.58.8...@times-components/author-profile@0.58.9) (2018-05-01)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-profile
+
+<a name="0.58.8"></a>
+## [0.58.8](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.58.7...@times-components/author-profile@0.58.8) (2018-04-30)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-profile
+
 <a name="0.58.7"></a>
 ## [0.58.7](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.58.6...@times-components/author-profile@0.58.7) (2018-04-30)