+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.7.7"></a>
+## [0.7.7](https://github.com/newsuk/times-components/compare/@times-components/icons@0.7.6...@times-components/icons@0.7.7) (2018-04-30)
+
+
+
+
+**Note:** Version bump only for package @times-components/icons
+
 <a name="0.7.6"></a>
 ## [0.7.6](https://github.com/newsuk/times-components/compare/@times-components/icons@0.7.5...@times-components/icons@0.7.6) (2018-04-30)